Better follow cam implementation
Added bounds to where the camera can pan. This used to keep the player
in the center while not going too far in any direction outside the
play area.

Had to somewhat jig the cam area size manually since my initial logic
was incorrect. Right now it kinda works how I wanted but am confused
as to why it works.